Refactor webpack and convert App to jsx
